warranty-claims-agent/requirements.txt
# LLM & Agent Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.2.0

# Structured Output & Validation
instructor>=1.0.0
pydantic>=2.0.0

# Document Processing (fallback to pdfplumber if docling fails)
pdfplumber>=0.10.0
# docling>=2.0.0  # Uncomment if using Python 3.10-3.12

# Vector Store & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# MCP (Model Context Protocol) - Optional
# mcp>=1.0.0  # Uncomment if needed

# Observability with Arize Phoenix (open source, free)
arize-phoenix>=8.0.0
openinference-instrumentation-langchain>=0.1.0
openinference-instrumentation-openai>=0.1.0

# Web UI
streamlit>=1.39.0

# PDF Generation
reportlab>=4.0.0

# File Watching
watchdog>=5.0.0

# Utilities
python-dotenv>=1.0.0
faker>=30.0.0
httpx>=0.27.0
aiosqlite>=0.20.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
